% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gm_clean.R
\name{gm_clean}
\alias{gm_clean}
\title{Clean source data.}
\usage{
gm_clean(geom)
}
\arguments{
\item{geom}{An sf object to clean.}
}
\value{
A cleaned sf object.
}
\description{
`gm_clean()` tests if all elements in the sf object are polygons and recasts
   them as MULTYPOLYGONs if necessary; validates the polygons; and reprojects
   the polygons to EPSG 6933, Global Equal Area.
}
\examples{
# Create a simple polygon as an sf object
library(sf)
geom <- st_as_sf(
  data.frame(
    id = 1,
    wkt = "POLYGON ((0 0, 0 1, 1 1, 1 0, 0 0))"
  ),
  wkt = "wkt",
  crs = 4326
)

gm_clean(geom)
}
