% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gm_global.r
\name{gm_global}
\alias{gm_global}
\title{Global summary statistics}
\usage{
gm_global(table, cutoff = 0.8, top.only = TRUE)
}
\arguments{
\item{table}{An object of class list created by [gm_crossmatch()]}

\item{cutoff}{Minimum areal similarity cut-off. Defaults to 0.8}

\item{top.only}{Only include top-ranked matches in the quantile table and histogram}
}
\value{
The function returns a list of tables:
  \itemize{
    \item top.only = TRUE if quantile table and histogram include only top-ranked matches
    \item quantiles = a `quantile()` summary of the similarity scores
    \item histogram = a histogram of the similarity scores with the cutoff indicated
    \item matched_cutoff1 = proportion of geom1 polygons matched relative to cutoff
    \item matched_cutoff2 = proportion of geom2 polygons matched relative to cutoff
    \item matched_toprank1 = proportion of geom1 polygons 'top rank' matched
    \item matched_toprank2 = proportion of geom2 polygons 'top rank' matched
    }
}
\description{
`gm_global()` returns a list of global summary statistics from the output of [gm_crossmatch()].
}
\examples{
library(geomatchR)

# Create example polygon set 1
sf1 <- st_sf(
  geoid = c("A1", "A2", "A3"),
  geoname1 = c("Alpha", "Beta", "Gamma"),
  geometry = st_sfc(
    st_polygon(list(rbind(c(0,0), c(0,2), c(2,2), c(2,0), c(0,0)))),
    st_polygon(list(rbind(c(2,0), c(2,2), c(4,2), c(4,0), c(2,0)))),
    st_polygon(list(rbind(c(0,2), c(0,4), c(2,4), c(2,2), c(0,2))))
  ), crs = 4326
)

# Create example polygon set 2
sf2 <- st_sf(
  geoid = c("B1", "B2", "B3"),
  geoname2 = c("Eins", "Zwei", "Drei"),
  geometry = st_sfc(
    st_polygon(list(rbind(c(0.25,0), c(0,2), c(2,2), c(2,0), c(0.25,0)))), # same
    st_polygon(list(rbind(c(2,0), c(2,2), c(4,2), c(4,0), c(2,0)))),
    st_polygon(list(rbind(c(0,2), c(0.5,4), c(2.25,4), c(2.75,2), c(0,2))))
  ), crs = 4326
)

# Run gm_crossmatch
result <- gm_crossmatch(
  geom1 = sf1,
  geom2 = sf2,
  id1 = "geoid",
  id2 = "geoid",
  name1 = "geoname1",
  name2 = "geoname2",
  cutoff = 0.5,
  clean = TRUE
)

# global summary statistics
global <- gm_global(
  table = result,
  cutoff = 0.5,
  top.only = TRUE
)

# inspect outputs
global$top.only
global$quantiles
global$histogram
global$matched_cutoff1
global$matched_cutoff2
global$matched_toprank1
global$matched_toprank2
}
