% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gm_map_similarity.R
\name{gm_map_similarity}
\alias{gm_map_similarity}
\title{Visually represent similarity scores}
\usage{
gm_map_similarity(
  geom1,
  geom2,
  id1,
  id2,
  name1,
  name2,
  table,
  label.cutoff = 0.8,
  labels = "none",
  plot = "combined"
)
}
\arguments{
\item{geom1}{The first polygon sf object}

\item{geom2}{The second polygon sf object}

\item{id1}{The unique polygon identification code of the first polygon sf object}

\item{id2}{The unique polygon identification code of the second polygon sf object}

\item{name1}{The unique polygon identification name of the first polygon sf object}

\item{name2}{The unique polygon identification name of the second polygon sf object}

\item{table}{An object of class list created by `gm_crossmatch()`}

\item{label.cutoff}{Minimum similarity score of polygons to label. Defaults to 0.8}

\item{labels}{"id" to show polygon identifiers defined by id1 and id2;
"name" to show polygon names defined by name1 and name2;
"none" to suppress polygon labels. Defaults to "none".}

\item{plot}{"combined" or "c" to create a two-panel plot showing forward and backward comparisons;
"forward" or "f" to show forward comparison only; "backward" or "b" to show backward comparison only.}
}
\value{
Returns a plot object
}
\description{
`gm_map_similarity()` creates a map of similarity scores.
}
\examples{
library(sf)

# Create example polygon set 1
sf1 <- st_sf(
  geoid = c("A1", "A2", "A3"),
  geoname1 = c("Alpha", "Beta", "Gamma"),
  geometry = st_sfc(
    st_polygon(list(rbind(c(0,0), c(0,2), c(2,2), c(2,0), c(0,0)))),
    st_polygon(list(rbind(c(2,0), c(2,2), c(4,2), c(4,0), c(2,0)))),
    st_polygon(list(rbind(c(0,2), c(0,4), c(2,4), c(2,2), c(0,2))))
  ), crs = 4326
)

# Create example polygon set 2
sf2 <- st_sf(
  geoid = c("B1", "B2", "B3"),
  geoname2 = c("Eins", "Zwei", "Drei"),
  geometry = st_sfc(
    st_polygon(list(rbind(c(0.25,0), c(0,2), c(2,2), c(2,0), c(0.25,0)))), # same
    st_polygon(list(rbind(c(2,0), c(2,2), c(4,2), c(4,0), c(2,0)))),
    st_polygon(list(rbind(c(0,2), c(0.5,4), c(2.25,4), c(2.75,2), c(0,2))))
  ), crs = 4326
)

# Run gm_crossmatch
result <- gm_crossmatch(
  geom1 = sf1,
  geom2 = sf2,
  id1 = "geoid",
  id2 = "geoid",
  name1 = "geoname1",
  name2 = "geoname2",
  cutoff = 0.5,
  clean = TRUE
)

# map similarity
gm_map_similarity(geom1 = sf1,
                  geom2 = sf2,
                  id1 = "geoid",
                  id2 = "geoid",
                  name1 = "geoname1",
                  name2 = "geoname2",
                  table = result,
                  label.cutoff = .5,
                  labels = "name",
                  plot = "combined"
)
}
